Rename some files to fit the model
