+++ war_or_peace_runner.rb
@@ -1,7 +1,7 @@
 require './lib/card'
 require './lib/deck'
 require './lib/player'
-require './lib/turn2'
+require './lib/turn'
 require './lib/card_generator'
 require 'pry'
